requirements.txt
certifi 
cycler==0.11.0
fonttools==4.38.0
imageio==2.31.1
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.2.2
networkx==2.6.3
numpy==1.21.4
opencv-python==4.2.0.32
packaging==23.1
Pillow==9.5.0
protobuf==4.23.4
pyparsing==3.1.0
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
tensorboardX==2.6.1
tifffile==2021.11.2
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.65.0
typing_extensions==4.7.1
